add aggreact-make-explicit-command
Motivated by transposition with explicit regions, to allow transpose keys to be repeatable when they start and end with a region active.  I like the current logic of the automatic repetition allowing repetition to capture the region creation process along with the final command on the region, and changing that to allow transposition repetition broke that.  I don't know that I love this design, but it works for what I want to accomplish and it's expedient.